scripts/data_collection/requirements.txt
requests>=2.31.0
tqdm>=4.66.0
pillow>=10.0.0
kaggle>=1.5.16
gdown>=4.7.1
beautifulsoup4>=4.12.0
selenium>=4.15.0
praw>=7.7.1
yt-dlp>=2023.10.13
opencv-python>=4.8.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
pandas>=2.1.0
scikit-learn>=1.3.0

